+++ lib/cocoapods/installer.rb
@@ -218,13 +218,7 @@ module Pod
     # @return [Array<PodTarget>] The model representations of pod targets
     #         generated as result of the analyzer.
     #
-    def pod_targets
-      aggregate_target_pod_targets = aggregate_targets.flat_map(&:pod_targets)
-      test_dependent_targets = aggregate_target_pod_targets.flat_map do |pod_target|
-        pod_target.test_dependent_targets_by_spec_name.values.flatten
-      end
-      (aggregate_target_pod_targets + test_dependent_targets).uniq
-    end
+    attr_reader :pod_targets
 
     # @return [Array<Specification>] The specifications that were installed.
     #
@@ -243,6 +237,7 @@ module Pod
     def analyze(analyzer = create_analyzer)
       @analysis_result = analyzer.analyze
       @aggregate_targets = @analysis_result.targets
+      @pod_targets = @analysis_result.pod_targets
     end
 
     def create_analyzer(plugin_sources = nil)
@@ -422,11 +417,9 @@ module Pod
     # @return [void]
     #
     def perform_post_install_actions
-      unlock_pod_sources
       run_plugins_post_install_hooks
       warn_for_deprecations
       warn_for_installed_script_phases
-      lock_pod_sources
       print_post_install_message
     end
 
@@ -444,8 +437,21 @@ module Pod
     # Runs the registered callbacks for the plugins post install hooks.
     #
     def run_plugins_post_install_hooks
-      context = PostInstallHooksContext.generate(sandbox, aggregate_targets)
-      HooksManager.run(:post_install, context, plugins)
+      # This short-circuits because unlocking pod sources is expensive
+      if any_plugin_post_install_hooks?
+        unlock_pod_sources
+
+        context = PostInstallHooksContext.generate(sandbox, aggregate_targets)
+        HooksManager.run(:post_install, context, plugins)
+      end
+
+      lock_pod_sources
+    end
+
+    # @return [Boolean] whether there are any plugin post-install hooks to run
+    #
+    def any_plugin_post_install_hooks?
+      HooksManager.hooks_to_run(:post_install, plugins).any?
     end
 
     # Runs the registered callbacks for the source provider plugin hooks.