+++ lib/cocoapods/target/aggregate_target.rb
@@ -204,7 +204,10 @@ module Pod
         framework_paths_by_config = {}
         user_build_configurations.keys.each do |config|
           relevant_pod_targets = pod_targets_for_build_configuration(config)
-          framework_paths_by_config[config] = relevant_pod_targets.flat_map { |pt| pt.framework_paths(false) }
+          framework_paths_by_config[config] = relevant_pod_targets.flat_map do |pod_target|
+            non_test_specs = pod_target.non_test_specs.map(&:name)
+            pod_target.framework_paths.values_at(*non_test_specs).flatten.compact.uniq
+          end
         end
         framework_paths_by_config
       end
@@ -219,7 +222,9 @@ module Pod
         end
         user_build_configurations.keys.each_with_object({}) do |config, resources_by_config|
           resources_by_config[config] = (relevant_pod_targets & pod_targets_for_build_configuration(config)).flat_map do |pod_target|
-            (pod_target.resource_paths(false) + [bridge_support_file].compact).uniq
+            non_test_specs = pod_target.non_test_specs.map(&:name)
+            resource_paths = pod_target.resource_paths.values_at(*non_test_specs).flatten.compact
+            (resource_paths + [bridge_support_file].compact).uniq
           end
         end
       end