+++ spec/fixtures/watermelon-lib/Classes/Watermelon.h
@@ -1,0 +1,6 @@
+
+#import <Foundation/Foundation.h>
+
+/** Watermelons are cool */
+@interface WatermelonObj : NSObject
+@end