+++ spec/unit/installer/xcode/pods_project_generator/pod_target_installer_spec.rb
@@ -100,6 +100,11 @@ module Pod
               end
             end
 
+            it 'cleans up temporary directories' do
+              @installer.expects(:clean_support_files_temp_dir).once
+              @installer.install!
+            end
+
             #--------------------------------------#
 
             describe 'headers folder paths' do
@@ -808,6 +813,14 @@ module Pod
               @installer.install!
               @project.targets.map(&:name).should == ['BananaLib']
               @project.targets.first.class.should == Xcodeproj::Project::PBXAggregateTarget
+            end
+
+            it 'adds xcconfig file reference for the aggregate placeholder native target' do
+              @pod_target.stubs(:should_build?).returns(false)
+              @installer.install!
+              @project.support_files_group
+              group = @project['Pods/BananaLib/Support Files']
+              group.children.map(&:display_name).sort.should == ['BananaLib.xcconfig']
             end
 
             #--------------------------------------------------------------------------------#